chore: remove dead orchestration provider replay helpers
