+++ docs/internal/baselines/deadcode-baseline.txt
@@ -83,25 +83,6 @@ pkg/services/work/transports/cli/work/list.go: unreachable func: List
 pkg/services/work/transports/cli/work/move.go: unreachable func: Move
 pkg/services/work/transports/cli/work/show.go: unreachable func: Show
 pkg/services/work/transports/cli/work/watch.go: unreachable func: RenderWatchTransition
-tests/functional/orchestration/javascript/durability/resume_test.go: unreachable func: resumeJavaScriptSessionExpectingInvalidState
-tests/functional/orchestration/javascript/events/events_test.go: unreachable func: decodeNDJSONRecords
-tests/functional/orchestration/javascript/events/events_test.go: unreachable func: mapKeys
-tests/functional/providers/acp/golden_fixture_test.go: unreachable func: readGoldenJSON
-tests/functional/providers/helpers_test.go: unreachable func: assertCursorProviderCompleted
-tests/functional/providers/helpers_test.go: unreachable func: assertProviderArgsPrompt
-tests/functional/providers/helpers_test.go: unreachable func: assertProviderStdin
-tests/functional/providers/helpers_test.go: unreachable func: cursorMergedPrompt
-tests/functional/providers/helpers_test.go: unreachable func: requireRuntimeLogPath
-tests/functional/replay_contracts/replay_process_test.go: unreachable func: assertReplayPlaceCounts
-tests/functional/replay_contracts/replay_process_test.go: unreachable func: observeReplayThroughRoot
-tests/functional/replay_contracts/replay_process_test.go: unreachable func: replayWorkIncludesID
-tests/functional/replay_contracts/replay_script_boundary_helpers_test.go: unreachable func: assertScriptEventsRecordedInArtifact
-tests/functional/replay_contracts/replay_script_boundary_helpers_test.go: unreachable func: normalizeReplayContractStdout
-tests/functional/replay_contracts/replay_script_boundary_helpers_test.go: unreachable func: requireScriptResponseEventIndices
-tests/functional/replay_contracts/short_helpers_test.go: unreachable func: assertReplayArtifactDoesNotContainRawValue
-tests/functional/replay_contracts/short_helpers_test.go: unreachable func: factoryRelationsValue
-tests/functional/replay_contracts/short_helpers_test.go: unreachable func: replayEventCount
-tests/functional/replay_contracts/short_helpers_test.go: unreachable func: strPtr
 tests/functional/sessions/execution/helpers_test.go: unreachable func: pollHostedCLIInvocationObservations
 tests/functional/sessions/execution/helpers_test.go: unreachable func: tryWaitForTerminalWorkPrimaryTextDuringInvocation
 tests/functional/sessions/execution/helpers_test.go: unreachable func: waitForTerminalWorkPrimaryText
