+++ tests/functional/replay_contracts/replay_process_test.go
@@ -1,3 +1,5 @@
+//go:build functionallong
+
 package replay_contracts
 
 import (
